+++ cmd/dReams/functions.go
@@ -201,16 +201,12 @@ func fetch(done chan struct{}) {
 					go menuRefresh(offset)
 
 					offset++
-					if offset >= 21 {
+					if offset >= 41 {
 						offset = 0
 					}
 				}
 
 				if rpc.Daemon.IsConnected() {
-					if rpc.Startup {
-						go menu.Info.RefreshPrice(App_Name)
-					}
-
 					rpc.Startup = false
 				}
 
@@ -248,7 +244,7 @@ func menuRefresh(offset int) {
 			menu.Info.SetStatus("Gnomon Syncing...")
 		}
 
-		if offset == 20 {
+		if offset == 40 || menu.Info.Price.Text == "" {
 			go menu.Info.RefreshPrice(App_Name)
 		}
 	}
@@ -506,16 +502,14 @@ func checkConnection() {
 		menu.Control.Check.Daemon.SetChecked(true)
 	} else {
 		menu.Control.Check.Daemon.SetChecked(false)
-		disableActions(true)
 		disconnected()
 	}
 
 	if rpc.Wallet.IsConnected() {
 		if rpc.Daemon.IsConnected() {
-			disableActions(false)
+			menu.Assets.Swap.Show()
 		}
 	} else {
-		disableActions(true)
 		disconnected()
 		gnomon.Checked(false)
 	}
@@ -526,21 +520,11 @@ func disconnected() {
 	holdero.Disconnected(menu.DappEnabled("Holdero"))
 	prediction.Disconnected()
 	rpc.Wallet.Address = ""
+	menu.Assets.Swap.Hide()
 	menu.Assets.Names.ClearSelected()
 	menu.Theme.Select.Options = menu.Control.Themes
 	menu.Theme.Select.Refresh()
 	menu.Assets.Asset = []menu.Asset{}
-}
-
-// Disable actions requiring connection
-func disableActions(d bool) {
-	if d {
-		menu.Assets.Swap.Hide()
-	} else {
-		menu.Assets.Swap.Show()
-	}
-
-	menu.Assets.Swap.Refresh()
 }
 
 // dReams search filters for Gnomon index
@@ -611,9 +595,14 @@ func gnomonFilters() (filter []string) {
 	}
 
 	if menu.DappEnabled("Grokked") {
-		grok := rpc.GetSCCode(grok.GROKSCID)
-		if grok != "" {
-			filter = append(filter, grok)
+		grokked := rpc.GetSCCode(grok.GROKSCID)
+		if grokked != "" {
+			filter = append(filter, grokked)
+		}
+
+		grokked = rpc.GetSCCode(grok.GROKOG)
+		if grokked != "" {
+			filter = append(filter, grokked)
 		}
 	}