+++ menu/prices.go
@@ -37,6 +37,22 @@ type kuFeed struct {
 	} `json:"data"`
 }
 
+type xeggexFeed struct {
+	TickerID       string `json:"ticker_id"`
+	Type           string `json:"type"`
+	BaseCurrency   string `json:"base_currency"`
+	TargetCurrency string `json:"target_currency"`
+	LastPrice      string `json:"last_price"`
+	BaseVolume     string `json:"base_volume"`
+	TargetVolume   string `json:"target_volume"`
+	UsdVolumeEst   string `json:"usd_volume_est"`
+	Bid            string `json:"bid"`
+	Ask            string `json:"ask"`
+	High           string `json:"high"`
+	Low            string `json:"low"`
+	ChangePercent  string `json:"change_percent"`
+}
+
 // Used for placing coin decimal, default returns 2 decimal place
 func CoinDecimal(ticker string) int {
 	split := strings.Split(ticker, "-")
@@ -54,56 +70,60 @@ func CoinDecimal(ticker string) int {
 	return 2
 }
 
-// Main price fetch, returns float and display string
-//   - Average from 3 feeds, if not take average from 2, if not TO value takes priority spot
+// Main price fetch, returns float and display string, average from 4 feeds
+//   - coin, "BTC-USDT", "DERO-USDT", "XMR-USDT", "DERO-BTC", "XMR-BTC"
 func GetPrice(coin, tag string) (price float64, display string) {
-	var t float64
-	var k float64
-	var g float64
+	var sum float64
+	var count int
 	priceT := getOgre(coin)
 	priceK := getKucoin(coin)
 	priceG := getGeko(coin)
+	priceX := getXeggex(coin)
 
 	if CoinDecimal(coin) == 8 {
 		if tf, err := strconv.ParseFloat(priceT, 64); err == nil {
-			t = tf * 100000000
+			sum += tf * 100000000
+			count++
 		}
 
 		if kf, err := strconv.ParseFloat(priceK, 64); err == nil {
-			k = kf * 100000000
+			sum += kf * 100000000
+			count++
 		}
 
 		if gf, err := strconv.ParseFloat(priceG, 64); err == nil {
-			g = gf * 100000000
+			sum += gf * 100000000
+			count++
+		}
+
+		if xf, err := strconv.ParseFloat(priceX, 64); err == nil {
+			sum += xf * 100000000
+			count++
 		}
 	} else {
 		if tf, err := strconv.ParseFloat(priceT, 64); err == nil {
-			t = tf * 100
+			sum += tf * 100
+			count++
 		}
 
 		if kf, err := strconv.ParseFloat(priceK, 64); err == nil {
-			k = kf * 100
+			sum += kf * 100
+			count++
 		}
 
 		if gf, err := strconv.ParseFloat(priceG, 64); err == nil {
-			g = gf * 100
+			sum += gf * 100
+			count++
+		}
+
+		if xf, err := strconv.ParseFloat(priceX, 64); err == nil {
+			sum += xf * 100
+			count++
 		}
 	}
 
-	if t > 0 && k > 0 && g > 0 {
-		price = (t + k + g) / 3
-	} else if t > 0 && k > 0 {
-		price = (t + k) / 2
-	} else if k > 0 && g > 0 {
-		price = (k + g) / 2
-	} else if t > 0 && g > 0 {
-		price = (t + g) / 2
-	} else if t > 0 {
-		price = t
-	} else if k > 0 {
-		price = k
-	} else if g > 0 {
-		price = g
+	if sum > 0 {
+		price = sum / float64(count)
 	} else {
 		price = 0
 		logger.Errorf("[%s] Error getting price feed\n", tag)
@@ -118,8 +138,8 @@ func GetPrice(coin, tag string) (price float64, display string) {
 	return
 }
 
-// Get TO coin price feed
-func getOgre(coin string) string {
+// Get TO pair feed
+func getOgre(coin string) (price string) {
 	decimal := 2
 	var url string
 	var found ogreFeed
@@ -137,34 +157,37 @@ func getOgre(coin string) string {
 		url = "https://tradeogre.com/api/v1/ticker/btc-xmr"
 		decimal = 8
 	default:
-		return ""
+		return
 	}
 
 	client := &http.Client{Timeout: 10 * time.Second}
 	req, err := http.NewRequest("GET", url, nil)
 	if err != nil {
 		logger.Errorln("[getOgre]", err)
-		return ""
+		return
 	}
 
 	req.Header.Add("Accept", "application/json")
 	req.Header.Add("Content-Type", "application/json")
+
 	resp, err := client.Do(req)
-
 	if err != nil {
 		logger.Errorln("[getOgre]", err)
-		return ""
+		return
 	}
-
 	defer resp.Body.Close()
-	b, err := io.ReadAll(resp.Body)
 
+	b, err := io.ReadAll(resp.Body)
 	if err != nil {
 		logger.Errorln("[getOgre]", err)
-		return ""
+		return
 	}
 
-	json.Unmarshal(b, &found)
+	err = json.Unmarshal(b, &found)
+	if err != nil {
+		logger.Errorln("[getOgre]", err)
+		return
+	}
 
 	if s, err := strconv.ParseFloat(found.Price, 64); err == nil {
 		if decimal == 8 {
@@ -176,8 +199,8 @@ func getOgre(coin string) string {
 	return found.Price
 }
 
-// Get Kucoin coin price feed
-func getKucoin(coin string) string {
+// Get Kucoin pair feed
+func getKucoin(coin string) (price string) {
 	decimal := 2
 	var url string
 	var found kuFeed
@@ -185,44 +208,49 @@ func getKucoin(coin string) string {
 	case "BTC-USDT":
 		url = "https://api.kucoin.com/api/v1/market/orderbook/level1?symbol=BTC-USDT"
 	case "DERO-USDT":
-		url = "https://api.kucoin.com/api/v1/market/orderbook/level1?symbol=DERO-USDT"
+		// url = "https://api.kucoin.com/api/v1/market/orderbook/level1?symbol=DERO-USDT"
+		return
 	case "XMR-USDT":
 		url = "https://api.kucoin.com/api/v1/market/orderbook/level1?symbol=XMR-USDT"
 	case "DERO-BTC":
-		url = "https://api.kucoin.com/api/v1/market/orderbook/level1?symbol=DERO-BTC"
-		decimal = 8
+		// url = "https://api.kucoin.com/api/v1/market/orderbook/level1?symbol=DERO-BTC"
+		// decimal = 8
+		return
 	case "XMR-BTC":
 		url = "https://api.kucoin.com/api/v1/market/orderbook/level1?symbol=XMR-BTC"
 		decimal = 8
 	default:
-		return ""
+		return
 	}
 
 	client := &http.Client{Timeout: 10 * time.Second}
 	req, err := http.NewRequest("GET", url, nil)
 	if err != nil {
 		logger.Errorln("[getKucoin]", err)
-		return ""
+		return
 	}
 
 	req.Header.Add("Accept", "application/json")
 	req.Header.Add("Content-Type", "application/json")
+
 	resp, err := client.Do(req)
-
 	if err != nil {
 		logger.Errorln("[getKucoin]", err)
-		return ""
+		return
 	}
-
 	defer resp.Body.Close()
-	b, err := io.ReadAll(resp.Body)
 
+	b, err := io.ReadAll(resp.Body)
 	if err != nil {
 		logger.Errorln("[getKucoin]", err)
-		return ""
+		return
 	}
 
-	json.Unmarshal(b, &found)
+	err = json.Unmarshal(b, &found)
+	if err != nil {
+		logger.Errorln("[getKucoin]", err)
+		return
+	}
 
 	if s, err := strconv.ParseFloat(found.Data.Price, 64); err == nil {
 		if decimal == 8 {
@@ -234,7 +262,7 @@ func getKucoin(coin string) string {
 	return found.Data.Price
 }
 
-// Get coingeko price feed
+// Get coingeko pair feed
 func getGeko(coin string) string {
 	client := &http.Client{Timeout: time.Second * 10}
 	CG := coingecko.NewClient(client)
@@ -269,4 +297,65 @@ func getGeko(coin string) string {
 	}
 
 	return fmt.Sprintf("%.2f", price.MarketPrice)
+}
+
+// Get Xeggex pair feed
+func getXeggex(coin string) (price string) {
+	decimal := 2
+	var url string
+	var found xeggexFeed
+	switch coin {
+	case "BTC-USDT":
+		url = "https://api.xeggex.com/api/v2/ticker/BTC/USDT"
+	case "DERO-USDT":
+		url = "https://api.xeggex.com/api/v2/ticker/DERO/USDT"
+	case "XMR-USDT":
+		url = "https://api.xeggex.com/api/v2/ticker/XMR/USDT"
+	case "DERO-BTC":
+		url = "https://api.xeggex.com/api/v2/ticker/DERO/BTC"
+		decimal = 8
+	case "XMR-BTC":
+		url = "https://api.xeggex.com/api/v2/ticker/XMR/BTC"
+		decimal = 8
+	default:
+		return
+	}
+
+	client := &http.Client{Timeout: 10 * time.Second}
+	req, err := http.NewRequest("GET", url, nil)
+	if err != nil {
+		logger.Errorln("[getXeggex]", err)
+		return
+	}
+
+	req.Header.Add("Accept", "application/json")
+	req.Header.Add("Content-Type", "application/json")
+
+	resp, err := client.Do(req)
+	if err != nil {
+		logger.Errorln("[getXeggex]", err)
+		return
+	}
+	defer resp.Body.Close()
+
+	b, err := io.ReadAll(resp.Body)
+	if err != nil {
+		logger.Errorln("[getXeggex]", err)
+		return
+	}
+
+	err = json.Unmarshal(b, &found)
+	if err != nil {
+		logger.Errorln("[getXeggex]", err)
+		return
+	}
+
+	if s, err := strconv.ParseFloat(found.LastPrice, 64); err == nil {
+		if decimal == 8 {
+			return fmt.Sprintf("%.8f", s)
+		}
+		return fmt.Sprintf("%.2f", s)
+	}
+
+	return found.LastPrice
 }